+++ src/bin/abci_node/data/leviathan_db/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000053
+MANIFEST-000061
